Representor issue:
on vdev device name should start with net_ionic keyword.
hence vdev=net_ionic0, net_ionic_pf0vf0 are applicable.

So when vdev is added, its name is stored on vdev bus and whenever
query comes for a device by its name, it would be searched on vdev bus.

Name to be stored is net_ionic_pf0vf0 and also ovs should pass the same
name , as in parsing, if we pass net_ionic, representor=pf0vf0, it will
fail because of ',' ..

OVS changes are required.. to pass net_ionic_pf0vf0
For simulator to get bar address also changes are done according to the name.